Handle ConnectionState.FAILED during login phase
This change fixes issue #129.

When a first login attempt fails, the connection state of the ConnectionService is set to FAILED, and it doesn't change. If a new login is performed, the doLogin method simply returns, as only the DISCONNECTED state is handled.

FAILED is now taken into account as well.